Add `combineMeasurementsControlsAndQuery`
Consolidate measurements controls and query param handling in
`combineMeasurementsControlsAndQuery`. Now that we load measurements
JSON upfront, we can use the collection's data to validate the query
params on initial load of the page. Handling all measurements
controls and query params in one place to make it easier to ensure
they are kept in sync.