Fix bug in narratives with filter changes
When we constructed the initial controls state we essentially did
```
controls.defaults = {
    filters: {}
},
controls.filters = controls.defaults.filters
```
and thus changing the current filter would also change the default. For
narratives we need to be able to reset the default state, and since each
value pointed to the same reference we ended up keeping the current
state.