Add tree-related types
Converted most but not all tree-related files. There are many type
errors exposed by conversion. They will be addressed in subsequent
commits.

I removed a few comments that are now self-explanatory with the added
types. I also moved all JSDoc types and descriptions into the function
signature.

Note that there are many optional properties in the newly defined
interfaces. This is not ideal and doesn't capture the fact that many
properties are conditional on things such as layout. But it's a starting
point to be improved over time.