Add types to measurements
Add types for most measurements specific files. I'm delaying converting
the measurementsD3.js file because the import of `event` from
d3-selection causes an error. This is likely due to the d3 types not
matching the version of d3-selection.¹

I decided to place the measurements JSON types in the same file as the
measurements state types so that it's easy to compare the data
structures before and after parsing.

There are many type errors that will be fixed in subsequent commits.

¹ <https://github.com/nextstrain/auspice/issues/1900>