Revert skipping of entropy calc on leading edge
This reverts commit 0c91724786d5b7f556defbcd002a9c003fac4962.

That commit was the first performance optimisation within #1879. The
idea was to skip the expensive recalculation when initially dragging the
date slider. In other contexts this change was a negative, specifically
when there was only one action (e.g. changing entropy to events) as we
essentially introduced a 500ms delay.

Note that when the panel is off-screen subsequent work in #1879 means
that the recalculation is skipped anyway, so reverting this commit only
affects performance when the panel is on screen and we're dragging the
date slider. To remedy this we should add a code path which specifically
skips the leading edge calculation when dragging the date slider.

Essentially closes #1899 - that issue was partly the recalculation
on-demand which remains, but greatly exacerbated by the unnecessary
500ms delay.