Update `Query` type
Update the `Query` type to reflect the returned type of
`querystring.parse`.¹

This led me to realize that there's a currently a bug in the measurement
filter query param matching when filtering using values that are _not_
a string the raw measurement metadata. This will be fixed in the
following commit.

¹ <https://nodejs.org/docs/latest-v22.x/api/querystring.html#querystringparsestr-sep-eq-options>