Convert all measurements metadata to string during parsing of JSON
Because the query params are always parsed as strings, the measurements
URL query `mf_*` will fail if the measurements' raw data are not strings.
Instead of doing string transformations throughout the code, just
convert all measurements metadata to strings during parsing of the JSON.

I've left the measurement `value` field as a number because converting
it to string resulted in a lot of calculation errors in D3. I will
revisit this in the future when I add types to the measurementsD3 file.

Renamed `MeasurementMetadata` to `JsonMeasurementMetadata` to
distinguish from `Measurement.[key: string]` as suggested by @victorlin
in review

<https://github.com/nextstrain/auspice/pull/1910#discussion_r1849130789>