Move comments to JSDoc
This allows some code editors to show the comment on hover.

I also added two new types (Phase, Frame) to replace some comments.

Notably, a type assertion in _frame is a necessary part of this
translation. I considered extracting the modulus operation into a
separate _mod3 function to reduce clutter in the _frame function,
however that seemed to be an unnecessary complication.