Add color tree by measurements feature
Add basic feature where clicking on the grouping value in the
measurements panel will color the tree tips by their average
measurement values. I decided to use the existing actions for adding
metadata/coloring and changing the color by so that the new measurements
will work as expected of normal colorings.

Many improvements on the feature to come in subsequent commits.