applyMeasurementsColorBy: Use multiple anchors for color scale
Using the default continuous color range to build the color scale.¹
Use quantile of all measurements as anchors for the continuous
measurements color scale to try to get an even distribution of colors
across the color range. In the future, we might want to expose color
scale as collection display parameters so that collections can specify
their own colors.

¹ <https://github.com/nextstrain/auspice/blob/c0ca42fed852bcd2709dfa853e134cd7ca6f2f46/src/util/colorScale.ts#L283>