measurements: Save color grouping value as a separate state
If the current `colorBy` state changes, then we lose the value of
the selected grouping for measurements colorings. Save this as a
separate control state so that it can be used to remove stale
measurements metadata and colorings.

This state is also used in the measurements panel to prevent extra
dispatches if the user clicks on the same grouping value multiple times.