wip: Update measurement coloring when changing collection
If the new collection has the same group by and grouping value that
is currently used for coloring, then just update the metadata and
color scale to match the new collection's measurements.

If the group by or grouping value is not valid for the new collection,
then remove the metadata and color scale from that was previously added.