Update measurement coloring when changing collection and group by
If there is a previous measurement grouping value, then remove the
measurement metadata and coloring that was previously added. If the
previous grouping value is the current color-by, then also set the
color-by back to the default color-by with a notification to the user.

If there is a new grouping value then add the measurement metadata and
coloring to match the latest measurements.