applyMeasurementsColorBy: remove old measurements coloring data
When applying a new measurement coloring, remove the old measurement
metadata and coloring from the state before adding the new data to
ensure that we only display the new coloring data.

This adds a new `REMOVE_METADATA` action type that is added to all the
reducers that update state for the `ADD_EXTRA_METADATA` action.