applyMeasurementsColorBy: remove old measurements coloring data
When applying a new measurement coloring, we don't want to keep the
previous measurement metadata and coloring, so dispatch an action to
remove the old data before adding the new measurement metadata and
coloring.

This adds a new `REMOVE_METADATA` action type that is added to all the
reducers that update state for the `ADD_EXTRA_METADATA` action.