Update measurements coloring when changing collection
Remove the measurement coloring if it's no longer valid for the new
collection. If the coloring is still valid, then update the metadata
and coloring to match the new measurements.

The `updateMeasurementsColorData` is a separate function because I will
be using the same logic in subsequent commits when changing group by
or filters.