[tree] allow display_defaults to define zoom label
This functionality was already possible via the url query 'label',
we now make this JSON-definable. Note that the "reset layout" button
still zooms out to view the entire tree(s).

One side effect with defining a display default label is that you can't
always specify a URL which will load the tree at the tree root. If
there's a branch label on the root node then you can almost get there by
"zooming" to the root node¹ which will then set a 'label' URL query.

Also note that when interacting with the tree, zooming to the same
label as the display default will set a URL query 'label', which differs
from our typical handling of defaults where we would not set any URL
query here.

¹ Not the `nodes[0]` root node, but the branch which appears as the
root node when the tree is rendered. This is why zooming to this node
appears slightly different from "reset layout".