Add URL query option for Focus on selected
This allows the state to be preserved in the URL and the ability to
focus on initial load.

The name 'focus' was chosen to match the name in state.

The value 'selected' was chosen to make room for upcoming features which
may also be under the focus banner, such as temporal focus.