Include link in "built with …" translation key
For some languages, it does not make sense to place the link at the end
of the sentence. The need to set placement per-language means it should
be included in the translation key, similar to existing keys formatted
as `{{...}}`.

This isn't possible with the `t()` function which only works with string
interpolation, but can be accomplished with `<Trans>` and a more verbose
translation key.

Replacements done in Visual Studio Code with regex search:

    ("Built with":\s*)"([^"]*)"(,?)

and replace:

    "Built with <Link>{{text}}</Link>": "$2 <Link>{{text}}</Link>"$3

The link has been included at the end to match existing behavior. It can
now be adjusted separately per language.