Focus on selected works in both dimensions
This causes date updates and other filters to dynamically update the XY
bounds, ensuring that the tree fills the page. This works well for
zooming to early samples and give a generally more dynamic feeling. It
combines very well with "focus on selected", which is why we combine
the two features.

Co-authored-by: james hadfield <jameshadfield@users.noreply.github.com>
Co-authored-by: Trevor Bedford <trevor@bedford.io>