Don't hide elements when streamtrees is on
Hiding these causes the toggle to jump up in the sidebar, which is
disorienting, especially if one wants to toggle between streamtrees
on/off for a quick visual comparison between the two.

Instead, disable them to indicate presence of the feature in other
scenarios. This aligns with behavior of the streamtrees toggle itself
when the feature is unsupported.