[streamtrees] don't hide sidebar layout options
Hiding these causes the toggle to jump up in the sidebar, which is
disorienting, especially if one wants to toggle between streamtrees
on/off for a quick visual comparison between the two.

Instead, disable them to indicate presence of the feature in other
scenarios. This aligns with behavior of the streamtrees toggle itself
when the feature is unsupported.

We use a 'not-allowed' cursor to indicate the button isn't available
following the guide of the existing sidebar toggles.

Co-authored-by: james hadfield <jameshadfield@users.noreply.github.com>