Zoom to selected is stored as a URL query
When clicking "Zoom to selected" we will attempt to add a URL query
`?treeZoom=selected` such that we can reload the tree to the same zoom
(subtree).

This behaviour is similar to when zooming into a node which has a
branch label which we store as a URL query (e.g. `?label=clade:A1`)
however the differences are instructive. While branch labels must
uniquely refer to a specific branch, the behaviour of "zoom to selected"
is conditional on the zoom level when the button was pressed - for a
full explanation of this see <https://github.com/nextstrain/auspice/pull/1321#issuecomment-2914923800>

The takeaway is that pressing "zoom to selected" will result in one
of three changes to the URL:
1. There is a branch label on the newly zoomed root node, so the URL
   query specifies `label=...`
2. We add `treeZoom=selected` as this exact view is recoverable by
   applying the filters to the entire dataset and then zooming to the
   selected subtree
3. Neither is added to the URL. This means the current filters would
   select tips which are currently off-screen, and thus the "zoom to
   selected" button's behaviour was contingent on the existing zoom
   level