Use SVG spinners
The React-controlled spinners now match the approach implemented for the
entry HTML in the previous commit. There's still a perceptible flash as
React unmounts the original spinner and replaces it with the new one and
the starting rotation resets.

(Actually, if we load an auspice dataset page directly there's now two
flashes as the rotation angle resets to zero... there's a flash as we
change from the entrypoint-HTML spinner to the React one within
`<MainComponentSwitch>`, which is described above, and there's a second
flash as the `<Main>` component loads and sets its own spinner, a flash
which was always there.)

The big improvement is that we no longer have a separate request for the
23KB PNG, which we remove entirely from Auspice. Rather we inline the
SVG (~4.4KB gzipped) which fixes the long-standing bug where placeholder
text would rotate rather than the logo, presumably because the image
hadn't arrived!