[bundles] separate out map component
The resulting bundle (`auspice.chunk.mapComponent.bundle.{hash}.js`) is
around 65Kb gzipped. When loading a dataset with the map toggled on
there's not a lot of improvement here, but for datasets without geo-
resolutions or with the map toggled off (via JSON or URL query) then
there are some nice improvements.