update polyfill approach
We already leveraged babel to rewrite JS syntax to match our 95%
browser target, so it makes sense to do the same for polyfilling.

The polyfill-focused bundle was ~60KB, which increaased to ~90KB when
we imported the entire (updated!) core-js library, and that decreases
to ~50KB when only including polyfills to reach our browser targets.