LinkOutModal: explicitly disable links for v1 datasets
As discussed in <https://github.com/nextstrain/nextstrain.org/issues/1188>,
the link out to v1 datasets never worked as expected because the nextstrain.org 
RESTful API does not support the v1 -> v2 conversion. Rather than try to build 
out the new feature for the old version, just explicitly disable to the links 
for v1 datasets as discussed in Slack
<https://bedfordlab.slack.com/archives/CBVHPFMGX/p1753470698246369>

This required adding `meta.extensions.original_version` in the v1 to v2 
conversion code to track the original version of datasets. This will require a 
matching change in the nextstrain.org Charon API.