[SVG download] Include measurements panel
Include measurements panel in the downloaded SVG, always displayed in the "full" 
mode so that all measurements hidden under the interactive scroll panel is 
displayed in the full height. This means the downloaded SVG does have a smaller 
"grid" tree panel that is not ideal, but I'm not wanting to pursue how to 
re-draw the tree for SVG downloads.

Note that the graph and the x-axis are separate SVGs in the measurements panel 
to allow the sticky x-axis, so they are parsed separate for the SVG download.