[SVG download] Fix map panel layout when "grid" display
Fix map panel layout when in the "Grid" display and the measurements panel is 
also displayed. Previously, the "full" map was added next to the "grid" tree in 
the downloaded SVG that was not ideal. This has bee an issue ever since the 
measurements panel was added because the measurements panel is prioritized over 
the map panel in the grid display.