Move escape key handling to modal component
This co-locates handling of the escape key within the component that it
is used for, where clearSelectedNode is already used as an onClick
callback to close the modal.

useEffect is the Hooks replacement¹ for componentDidMount and
componentWillUnmount. selectedNodeRef has been added to replicate the
behavior of adding the event listener once on component render. Without
the ref, selectedNode would need to be set as a dependency, which would
add/remove the event listener on every change of selectedNode.

¹ https://legacy.reactjs.org/docs/hooks-effect.html#effects-with-cleanup