[frequencies] use legend display as label
For colorings which have a JSON-defined legend block we now use the
"display" values rather then displaying the bounds. This  replicates the
behaviour in the tree/map legend. (Previously we would show the bounds.)

There is an edge case where (any number of) legend entries may not have
the "display" key set (it's optional). In this case we show the
(required) "value" key's value. In some ways this is worse than showing
the bounds, but it mirrors what we do in the tree/map legend.