show/update colorscale bounds
The bounds (for a legend entry) were previously only displayed in the
frequencies panel. Here we make two changes:

Firstly, for the frequencies panel update the formatting to use
mathematical notation to indicate that they're exclusive of the lower
bound, inclusive of the upper.

Secondly, in the (rare!) situation where a JSON colorscale defined the
`legend` block and included the (optional) bounds for each entry but
didn't provide a `display` value, then display the bounds.

I considered displaying the bounds for _all_ continuous scales but when
the bounds are chosen by Auspice it's less ideal. (Still, maybe we
should show them in a pop-up or similar.)