WIP color scale domain includes legend anchors
Improves the situation where the JSON defines legend anchors but not a
scale by forcing the color-scale domain to include said anchors.

The implementation is ok, but the legend/anchors block of the JSON scale
isn't yet validated when we create the scale domain so we are using
unvalidated data. Improving this would involve a bigger refactor which
doesn't seem worth it at the moment.

Needs more testing!