Tree buttons for each tree
For single trees (i.e. 99% of Auspice usage) there should be no changes.

For tangle-trees I often want more fine-grained control than the current
situation and making the buttons per-tree gives me this. There are
occasions where we'll now have to click two buttons, but that's an
acceptable trade off for me. (Perhaps we can implement a shift-click to
apply to both trees if this really frustrates someone!)

Aesthetically we need a little more space above trees to duplicate the
buttons and also have the tree names shown. The tree names themselves
were implemented as an afterthought really and could be improved one day.